Rename module: molecule --> io
