+++ example-py/example_2.py
@@ -19,7 +19,7 @@ coor = [
 # If the first atom is not metal atom, you can rearrange the sequence
 # of atom in list using coord.extract_octa method.
 
-atom_octa, coord_octa = oc.molecule.extract_octa(atom, coor)
+atom_octa, coord_octa = oc.io.extract_octa(atom, coor)
 
 dist = oc.CalcDistortion(coord_octa)
 zeta = dist.zeta  # Zeta
